requirements.txt
torch>=2.2
transformers>=4.40
modelscope>=1.15
openai>=1.0
tqdm>=4.66
numpy>=1.24
Pillow>=10.0
shortuuid>=1.0
decord>=0.6
tiktoken>=0.6
safetensors>=0.4
pycocotools>=2.0
matplotlib>=3.8
scipy>=1.11
scikit-learn>=1.3
spacy>=3.7
bert-score>=0.3
nltk>=3.8
llava @ git+https://github.com/haotian-liu/LLaVA.git

